explore page bug fix & code refactoring
